Move expect/actual UI helpers to core.ui.platform subpackage
Aligns with the *.platform.* package rule for expect/actual declarations.
Moved HoverAware, SystemBars, SystemTheme from core.ui/core.ui.theme to
core.ui.platform/core.ui.platform.theme across all 5 source sets.
Documented Room MemoDatabaseConstructor as a same-package exception.
